\name{hwang-package}
\alias{hwang-package}
\alias{hwang}
\docType{package}
\title{
\packageTitle{hwang}
}
\description{
\packageDescription{hwang}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{hwang}
\packageIndices{hwang}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{hwang}

Maintainer: \packageMaintainer{hwang}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
